backdoor/requirements.txt
Brotli==1.0.9
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
colorlog==6.7.0
Cython==0.29.34
EasyProcess==1.1
entrypoint2==1.1
idna==3.4
imutils==0.5.4
keyboard==0.13.5
MouseInfo==0.1.3
mss==9.0.1
mutagen==1.46.0
numpy==1.24.2
opencv-python==4.7.0.72
Pillow==9.5.0
pycryptodomex==3.17
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.28
python-dotenv==1.0.0
pytweening==1.0.4
pyzmq==24.0.1
requests==2.28.2
simplejpeg==1.6.6
six==1.16.0
tqdm==4.65.0
urllib3==1.26.15
vidgear==0.3.0
websockets==11.0.2
yt-dlp==2023.3.4
